Add files via upload
Results
After analyzing data of 212331 rows, the District Borough Number (DBN) reported 1018 tested.
This is followed by the summary below.  

From the summary statistics output displayed by DBN school below
Min.  1st Qu.   Median    Mean      3rd Qu.       Max. 
 1.00   5.00       20.0         49.36      57.00          1729.00

We see that the mean is 49.36 median numbers is 20.00; the measures of central tendency are close. 
The Percent Scoring 80 or above had 1000 unique values from the total of 212331
rows with top being s and a frequency of 7522. The Percent Scoring below 65 had 992 unique values from 212331 rows with the supreme beings and a frequency of 7522. The School Name had 1012 unique values from 212331 rows, with the top being New Visions Charter High School for Advanced Math and a frequency of 1337.